feat(mobile): add camera permission handling component and integrate it into QR scanner modal
